+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 ### Added
 
+## [102.0.4] - 2020-05-02
+### Added
+- csv and excel exports
+- show custome forms in adminhtml menu
+- permissions to view and save custom forms
+
 ## [102.0.3] - 2020-04-14
 ### Fixed
 - fixed issue with missing created_at column on fresh installation